Rename `anycubic_model_cloud_upload`
